+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 - **httperr:** Moved HTTP related code to separate package
+- **bugsnag:** Improve logger
 
 
 ## [0.9.1] - 2018-07-27